+++ src/main/java/org/dbflute/intro/app/logic/dfprop/database/DatabaseInfoLogic.java
@@ -13,15 +13,16 @@
  * either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-package org.dbflute.intro.app.logic.database;
+package org.dbflute.intro.app.logic.dfprop.database;
+
+import javax.annotation.Resource;
 
 import org.dbflute.intro.dbflute.allcommon.CDef;
 import org.dbflute.intro.dbflute.exbhv.ClsTargetDatabaseBhv;
 
-import javax.annotation.Resource;
-
 /**
  * @author ryohei
+ * @author jflute
  */
 public class DatabaseInfoLogic {
 
@@ -31,10 +32,11 @@ public class DatabaseInfoLogic {
     public boolean isEmbeddedJar(CDef.TargetDatabase target) {
         // done hakiba pri.B orElseTranslatingThrow() is better by jflute (2017/04/27)
         if (target == null) {
+            // #thinking jflute can it be not null? (should be strict) (2021/04/18)
             return false;
         }
         return databaseBhv.selectEntity(cb -> cb.query().setDatabaseCode_Equal_AsTargetDatabase(target))
-            .map(database -> database.isEmbeddedJarFlgTrue())
-            .orElseTranslatingThrow(cause -> new IllegalStateException("not found target database:" + target.alias(), cause));
+                .map(database -> database.isEmbeddedJarFlgTrue())
+                .orElseTranslatingThrow(cause -> new IllegalStateException("not found target database:" + target.alias(), cause));
     }
 }
